pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-detect"
version = "2.0.0"
description = "CV defect detection with deep learning anomaly models"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0,<1.0.0",
    "numpy>=1.24.0,<2.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "Pillow>=10.0.0,<12.0.0",
    "matplotlib>=3.7.0,<4.0.0",
    "scipy>=1.11.0,<2.0.0",
    "scikit-image>=0.21.0,<1.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn>=0.23.0,<1.0.0",
    "python-multipart>=0.0.6,<1.0.0",
]
onnx = ["onnxruntime>=1.15.0"]
ocr = ["pytesseract>=0.3.10", "paddleocr>=2.7.0"]
barcode = ["pyzbar>=0.1.9"]
camera = ["harvesters>=1.4.0"]
dev = ["pytest>=7.0", "ruff>=0.4.0", "pyinstaller>=6.0"]

[tool.setuptools.packages.find]
include = ["dl_anomaly*", "variation_model*", "shared*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
